Add the same compile flags as with ros2_controllers and fix errors (#185)

+++ include/realtime_tools/realtime_helpers.hpp
@@ -88,7 +88,7 @@ std::pair<bool, std::string> set_current_thread_affinity(int core);
  * \ref https://stackoverflow.com/a/150971
  * \returns The number of processors currently online (available)
 */
-int get_number_of_available_processors();
+int64_t get_number_of_available_processors();
 
 }  // namespace realtime_tools
